+++ app/assets/javascripts/administrate/components/has_many_form.js
@@ -1,3 +1,3 @@
 $(function() {
-  $(".form-field--has-many select").selectize({});
+  $(".field-unit--has-many select").selectize({});
 });