+++ lib/generators/administrate/assets/assets_generator.rb
@@ -1,0 +1,13 @@
+require "administrate/view_generator"
+
+module Administrate
+  module Generators
+    class AssetsGenerator < Administrate::ViewGenerator
+      def copy_assets
+        call_generator("administrate:assets:images")
+        call_generator("administrate:assets:javascripts")
+        call_generator("administrate:assets:stylesheets")
+      end
+    end
+  end
+end