+++ spec/administrate/views/fields/polymorphic/_index_spec.rb
@@ -26,7 +26,7 @@ describe "fields/polymorphic/_index", type: :view do
 
       render(
         partial: "fields/polymorphic/index.html.erb",
-        locals: { field: polymorphic },
+        locals: { field: polymorphic, namespace: "admin" },
       )
 
       expected = "<a href=\"#{product_path}\">#{product.name}</a>"