+++ spec/generators/views_generator_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require "rails_helper"
 require "generators/administrate/views/views_generator"
 require "support/generator_spec_helpers"
 
@@ -11,9 +11,22 @@ describe Administrate::Generators::ViewsGenerator, :generator do
       run_generator [resource]
 
       %w[index show new edit].each do |generator|
-        expect(Rails::Generators).to have_received(:invoke).
-          with("administrate:views:#{generator}", [resource])
+        expect(Rails::Generators).
+          to invoke_generator("administrate:views:#{generator}", [resource])
       end
+    end
+
+    it "revokes sub-generators if run through `rails destroy`" do
+      allow(Rails::Generators).to receive(:invoke)
+      resource = "users"
+
+      run_generator [resource], behavior: :revoke
+
+      expect(Rails::Generators).to invoke_generator(
+        "administrate:views:index",
+        [resource],
+        behavior: :revoke,
+      )
     end
   end
 end