+++ spec/rails_helper.rb
@@ -24,6 +24,10 @@ RSpec.configure do |config|
   config.infer_base_class_for_anonymous_controllers = false
   config.infer_spec_type_from_file_location!
   config.use_transactional_fixtures = false
+
+  config.before(:each, type: :generator) do
+    allow(Rails).to receive(:root).and_return(Pathname.new(file(".")))
+  end
 end
 
 ActiveRecord::Migration.maintain_test_schema!